processing-server/requirements.txt
requests
python-dotenv
supabase
fastapi
uvicorn[standard]
watchdog
Pillow
weasyprint
httpx
